move waas scraper to designated folder
